Change RDBG_FTL and RDBG_MNT macros.
Now they return anyway (even if compiled without their flags).
This is both logical - the print should be only when debugging, but flow
should be same, and compatible with current code - it allows to use them
in more places.
Also change THREAD_SERVER to return if either socket or listen fail.

Change-Id: I1ea482f8123c8592085543108ed43a375d5443ab