Update patch set 2
Patch Set 2: -Verified

Eyal, I have merged the object pass through patch and moved tests there. The only change we need from this patch is a one from .gitignore which is missing in my patch. In fact, this example demonstrates the benefit of small patches and CR :-)

Patch-set: 2
Label: Verified=0