Update patch set 2
Patch Set 2:

I know what reverting is, but reverting this to exactly as it was before also means return ssd.h and ssd.c Nir has deleted, and probably undoing more things in the object-pass through patch.
When the commit was introduced I said that it contains many unrelated changes and asked you to separate them to different patches.
I think this is the least bad way to handle it.
Ok, lets discuss the pros and cons.

Patch-set: 2