Combine mkfs.ext4 and nvme driver test
a new test suite in pytest which makes an ext4 filesystem
on the nvme device in the guest.
It then preformes rmmod nvme, insmod nvme, mounts the fs and generates
random files in various sizes on the mounted fs.
The test suite continues with unmounting, rmmod nvme, insmod nvme,
remounting and diffing file contents 10 times. Eventually it cleans
up and unmounts the device.

Change-Id: Id4be9fe9ab684b5a33903e9c6a377c01b6581dce