Integrating the monitor with the new logger
Currently, The monitor uses sockets in order to get the logs from
the simulator.
This patch changes the monitor so that it will support the new
logging mechanism implemented in eVSSIM/LOG_MGR, via the code in
monitor_test.h/cpp.
This patch also adds all the needed log types for the analyzer.

Change-Id: I0d0830eee9ba5a44d4482dc0d4c0048fbc01a1ac