[WIP] Adding logging server
Currently, the monitor is written in an outdated QT version, and
displays the statistics to the user locally.
This patch adds logging_server: a server which runs as a thread,
and contains both an http server and a web socket server. Requests
to the HTTP server will return a JS landing page, which connects to
the server using a websocket and displays up to date statistics.

Change-Id: Ib94a02bfc502a53d024417c7e9070e1121f1628e