Update patch set 3
Patch Set 3:

Thanks for working on this. I took a look on the changes you did in 441266. 
Please write tests as commented here and let's continue to work on this here. 
Please replace tabs with spaces in your editor ( not the first time I mentioned it that it makes comparison annoying)

Patch-set: 3