Update docker image to ubuntu 24.04
This update maintains our current use of GCC 4.6, with plans to transition to the
latest GCC version in the future, which will require extensive changes to the
codebase.

By moving to Ubuntu 24.04, we aim to leverage the latest and most advanced
tools available.

Depends-On: https://review.gerrithub.io/c/davidsaOpenu/kernel/+/1199903
Depends-On: https://review.gerrithub.io/c/davidsaOpenu/nvme-cli/+/1199904
Depends-On: https://review.gerrithub.io/c/davidsaOpenu/tnvme/+/1199905

Trello-Card: https://trello.com/c/4XskZmBC/113-update-docker-image-to-ubuntu-2404

Change-Id: I2cfd63138907dc571445b40b4103645f3cfe5483