Initial integration of NVMe, replacing OSD
This change modifies the script that format and mount a new exoFS
image. The script uses NVMe device instead of OSD device.
This patch begins the transition of exofs from using OSD devices to NVMe
devices for mounting and file operations, This change adds the implementation of the mount operation,
providing the starting point for the complete implementation of exofs over NVMe.

Co-authored-by: Amit Shoham <amitshoham8950@gmail.com>, Rami Krel <rlmk1997@gmail.com>, Adam Zats <adamrnd2@gmail.com>

Trello: https://trello.com/c/Gt8GxnCb/102-format-and-mount-a-new-exofs-image-using-nvme-device-instead-of-osd-device

Depends-On: https://review.gerrithub.io/c/davidsaOpenu/kernel/+/1200079
Depends-On: https://review.gerrithub.io/c/davidsaOpenu/open-osd/+/1200073

Change-Id: I904eef4f5208842e742def14704ab4c29e2672f3