Replace foreground GC implementation
Update _FTL_WRITE_SECT to utilize VICTIM_OVERALL_GC if VICTIM_OVERALL
fails, and perform foreground GC in such case. This allows to overwrite
sectors even when the disk is completely full, thanks to the new zero
pages reserve.

Fix an issue with DEFAULT_NEXT_PAGE_ALGO/GET_EMPTY_BLOCK where an empty
block is still considered an empty block, even though it's already full.
Now istead of waitig to the next write, I mark it as a victim block
right away. That was necessary for this story.

https://trello.com/c/WwcgFx5d/47-replace-foreground-gc-implementation


Change-Id: I361b145284d278c4c9b0bc85deb2a9a76115bb80